split queue and fix imports
